# test optimizations