--- a/crates/pagecache/src/snapshot.rs
+++ b/crates/pagecache/src/snapshot.rs
@@ -265,7 +265,10 @@ fn write_snapshot(config: &Config, snapshot: &Snapshot) -> Result<()> {
     f.write_all(&len_bytes)?;
     maybe_fail!("snap write crc");
     f.write_all(&crc32)?;
-    f.sync_all()?;
+
+    if !config.temporary {
+        f.sync_all()?;
+    }
     maybe_fail!("snap write post");
 
     trace!("wrote snapshot to {}", path_1.to_string_lossy());