--- a/tests/tests/test_log.rs
+++ b/tests/tests/test_log.rs
@@ -676,7 +676,7 @@ fn prop_log_works(ops: Vec<Op>, flusher: bool) -> bool {
 fn quickcheck_log_works() {
     QuickCheck::new()
         .gen(StdGen::new(rand::thread_rng(), 100))
-        .tests(1000)
+        .tests(100)
         .max_tests(10000)
         .quickcheck(prop_log_works as fn(Vec<Op>, bool) -> bool);
 }