these commits seek to reduce overall complexity and reduce some latency during system initialization that corresponds to poor test and benchmarking performance when hundreds or thousands of instances need to be created quickly