@mxinden 2019-07-21
Cool, that was quick. Ran it against my not-sophisticated criterion benchmark.

Before:
```
pagecache profile:                                                                                                                                                                                                   
               op |   min (us) |   med (us) |    90 (us) |    99 (us) |  99.9 (us) | 99.99 (us) |   max (us) |      count |    sum (s)                                                                               
--------------------------------------------------------------------------------------------------------------------------------------                                                                               
tree:                                                                                                                                                                                                                
            start |     4592.8 |    12115.5 |    65660.0 |    90422.2 |   149080.9 |   178482.3 |   178482.3 |       5177 |    116.110                                                                               
              set |        2.5 |        4.2 |       14.3 |       33.5 |       83.3 |      101.7 |    14076.3 |    5177000 |     34.890                                                                               
         traverse |        0.7 |        1.4 |        3.8 |       18.8 |       53.1 |       80.8 |      585.4 |   10354000 |     23.465                                                                               
              get |        1.1 |        1.7 |        3.0 |        6.8 |       17.3 |       36.3 |      305.6 |    5177000 |     10.480                                                                               
            merge |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |          0 |      0.000                                                                               
              del |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |          0 |      0.000                                                                               
              cas |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |          0 |      0.000                                                                               
             scan |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |          0 |      0.000                                                                               
         rev scan |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |          0 |      0.000                                                                               
tree contention loops: 0                                                                                                                                                                                             
tree split success rates: child(176018/176018) parent(176018/176018) root(0/0)
```

After:
```
pagecache profile:
               op |   min (us) |   med (us) |    90 (us) |    99 (us) |  99.9 (us) | 99.99 (us) |   max (us) |      count |    sum (s)
--------------------------------------------------------------------------------------------------------------------------------------
tree:
            start |     3471.2 |     9065.6 |    58235.2 |    66319.9 |   111551.8 |   133551.8 |   133551.8 |       5177 |     94.389
              set |        2.5 |        3.9 |       10.2 |       25.8 |       80.8 |       94.8 |     4925.8 |    5177000 |     30.534
         traverse |        0.7 |        1.3 |        2.9 |       16.2 |       37.0 |       76.1 |      722.2 |   10354000 |     20.946
              get |        1.1 |        1.6 |        2.5 |        6.1 |       17.2 |       31.6 |      221.9 |    5177000 |      9.604
            merge |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |          0 |      0.000
              del |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |          0 |      0.000
              cas |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |          0 |      0.000
             scan |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |          0 |      0.000
         rev scan |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |        NaN |          0 |      0.000
tree contention loops: 0
tree split success rates: child(176018/176018) parent(176018/176018) root(0/0)
```