@spacejam 2019-07-21
@mxinden still a ways to go but now a few of the offenders are visible!